requirements.txt
telethon==1.36.0
python-dotenv==1.0.1
PyYAML==6.0.2
rapidfuzz==3.9.6
